[eventjournaler] [LINUXDAR-3142] Changed: Removed unused exception, fixed multiple import errors, changed reader logging to debug (for now), and fixed queue popper namespace. Added: Test for telemetry, and a test for full queue implementation.

+++ eventjournaler/modules/SubscriberLib/EventQueuePusher.cpp
@@ -21,7 +21,7 @@ namespace SubscriberLib
         auto& telemetryHelper = Common::Telemetry::TelemetryHelper::getInstance();
         if (!m_eventQueue->push(event))
         {
-            telemetryHelper.increment(Plugin::Telemetry::telemetryDroppedAvEvents, ulong(1));
+            telemetryHelper.increment(Plugin::Telemetry::telemetryDroppedAvEvents, 1L);
         }
     }
 }
